test: address coverage review feedback
