+++ src/repo.rs
@@ -141,40 +141,5 @@ fn get_current_branch_name(repo: &Repository) -> Result<String, git2::Error> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use git2::{Oid, Signature};
-    use tempfile::tempdir;
-
-    fn create_commit(repo: &Repository) -> Oid {
-        let tree_id = repo.index().unwrap().write_tree().unwrap();
-        let tree = repo.find_tree(tree_id).unwrap();
-        let signature = Signature::now("Test", "test@example.com").unwrap();
-        repo.commit(Some("HEAD"), &signature, &signature, "test", &tree, &[])
-            .unwrap()
-    }
-
-    #[test]
-    fn test_current_branch_name() {
-        let temp_dir = tempdir().unwrap();
-        let repo = Repository::init(temp_dir.path()).unwrap();
-        repo.set_head("refs/heads/test-branch").unwrap();
-        create_commit(&repo);
-
-        assert_eq!(get_current_branch_name(&repo).unwrap(), "test-branch");
-    }
-
-    #[test]
-    fn test_detached_head_name() {
-        let temp_dir = tempdir().unwrap();
-        let repo = Repository::init(temp_dir.path()).unwrap();
-        let commit_id = create_commit(&repo);
-        repo.set_head_detached(commit_id).unwrap();
-
-        assert_eq!(get_current_branch_name(&repo).unwrap(), "detached HEAD");
-    }
-}
-
-#[cfg(test)]
 #[path = "../tests/crate/repo.rs"]
-mod clone_error_tests;
+mod tests;
